+++ app/jobs/podcast_episodes/create_job.rb
@@ -3,7 +3,7 @@ module PodcastEpisodes
     queue_as :podcast_episode_create
 
     # @param podcast_id [Integer] - podcast id
-    # @param item [Hash] according to the Podcasts::Episodes::RssItemData::ATTRIBUTES
+    # @param item [Hash] according to the Podcasts::EpisodeRssItem::ATTRIBUTES
     def perform(podcast_id, item)
       Podcasts::CreateEpisode.call(podcast_id, item)
     end
