+++ spec/services/podcasts/create_episode_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Podcasts::CreateEpisode, type: :service do
 
   context "when item has an https media_url" do
     let(:rss_item) { RSS::Parser.parse("spec/support/fixtures/podcasts/developertea.rss", false).items.first }
-    let!(:item) { Podcasts::Episodes::RssItemData.from_item(rss_item) }
+    let!(:item) { Podcasts::EpisodeRssItem.from_item(rss_item) }
 
     before do
       stub_request(:head, item.enclosure_url).to_return(status: 200)
@@ -37,7 +37,7 @@ RSpec.describe Podcasts::CreateEpisode, type: :service do
 
   context "when item has an http media url" do
     let(:rss_item) { RSS::Parser.parse("spec/support/fixtures/podcasts/awayfromthekeyboard.rss", false).items.first }
-    let!(:item) { Podcasts::Episodes::RssItemData.from_item(rss_item) }
+    let!(:item) { Podcasts::EpisodeRssItem.from_item(rss_item) }
     let(:https_url) { "https://awayfromthekeyboard.com/wp-content/uploads/2018/02/Episode_075_Lara_Hogan_Demystifies_Public_Speaking.mp3" }
 
     it "sets media_url to https version when it is available" do
